requirements.txt
python-telegram-bot
transformers
torch
python-dotenv